pyproject.toml
[tool.poetry]
name = "etlas api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
